requirements.txt
ultralytics>=8.0.0
supervision>=0.18.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
